+++ src/main/java/com/invadermonky/hungrypouches/proxy/CommonProxy.java
@@ -3,6 +3,7 @@ package com.invadermonky.hungrypouches.proxy;
 import com.invadermonky.hungrypouches.HungryPouches;
 import com.invadermonky.hungrypouches.client.GuiHandlerHP;
 import com.invadermonky.hungrypouches.events.EventItemPickup;
+import com.invadermonky.hungrypouches.network.PacketHandlerHP;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
@@ -11,6 +12,7 @@ import net.minecraftforge.fml.common.network.NetworkRegistry;
 
 public class CommonProxy {
     public void preInit(FMLPreInitializationEvent event) {
+        PacketHandlerHP.registerMessages();
         MinecraftForge.EVENT_BUS.register(EventItemPickup.INSTANCE);
         NetworkRegistry.INSTANCE.registerGuiHandler(HungryPouches.INSTANCE, new GuiHandlerHP());
     }